+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkObject.cs
@@ -821,6 +821,13 @@ namespace Unity.Netcode
                     Debug.LogWarning($"{ChildNetworkBehaviours[i].gameObject.name} is disabled! Netcode for GameObjects does not support spawning disabled NetworkBehaviours! The {ChildNetworkBehaviours[i].GetType().Name} component was skipped during spawn!");
                 }
             }
+            for (int i = 0; i < ChildNetworkBehaviours.Count; i++)
+            {
+                if (ChildNetworkBehaviours[i].gameObject.activeInHierarchy)
+                {
+                    ChildNetworkBehaviours[i].VisibleOnNetworkSpawn();
+                }
+            }
         }
 
         internal void InvokeBehaviourNetworkDespawn()