+++ com.unity.netcode.gameobjects/Tests/Runtime/DeferredMessagingTests.cs
@@ -198,29 +198,69 @@ namespace Unity.Netcode.RuntimeTests
         {
             ClientRpcCalled = true;
         }
+
+        public static readonly List<ulong> ClientInstances = new List<ulong>();
+        public override void OnNetworkSpawn()
+        {
+            if (!IsServer)
+            {
+                ClientInstances.Add(NetworkManager.LocalClientId);
+            }
+            base.OnNetworkSpawn();
+        }
     }
 
     public class DeferredMessageTestNetworkVariableComponent : NetworkBehaviour
     {
-        public NetworkVariable<int> TestNetworkVariable = new NetworkVariable<int>();
+        public static readonly List<ulong> ClientInstances = new List<ulong>();
+
+        public NetworkVariable<int> TestNetworkVariable;
+
+        public void Awake()
+        {
+            TestNetworkVariable = new NetworkVariable<int>(default, NetworkVariableReadPermission.Everyone, NetworkVariableWritePermission.Owner);
+        }
+
+        public override void OnNetworkSpawn()
+        {
+            if (!IsServer)
+            {
+                ClientInstances.Add(NetworkManager.LocalClientId);
+            }
+            base.OnNetworkSpawn();
+        }
     }
 
     public class DeferredMessageTestRpcAndNetworkVariableComponent : NetworkBehaviour
     {
+        public static readonly List<ulong> ClientInstances = new List<ulong>();
         public bool ClientRpcCalled;
+        public NetworkVariable<int> TestNetworkVariable;
+
+        public void Awake()
+        {
+            TestNetworkVariable = new NetworkVariable<int>(default, NetworkVariableReadPermission.Everyone, NetworkVariableWritePermission.Owner);
+        }
+
+        public override void OnNetworkSpawn()
+        {
+            if (!IsServer)
+            {
+                ClientInstances.Add(NetworkManager.LocalClientId);
+            }
+            base.OnNetworkSpawn();
+        }
 
         [ClientRpc]
         public void SendTestClientRpc()
         {
             ClientRpcCalled = true;
         }
-
-        public NetworkVariable<int> TestNetworkVariable = new NetworkVariable<int>();
     }
 
     public class DeferredMessagingTest : NetcodeIntegrationTest
     {
-        protected override int NumberOfClients => 2;
+        protected override int NumberOfClients => 0;
 
         private List<SpawnCatcher> m_ClientSpawnCatchers = new List<SpawnCatcher>();
 
@@ -228,31 +268,16 @@ namespace Unity.Netcode.RuntimeTests
         private GameObject m_NetworkVariablePrefab;
         private GameObject m_RpcAndNetworkVariablePrefab;
 
+        private int m_NumberOfClientsToLateJoin = 2;
+
         protected override IEnumerator OnSetup()
         {
+            DeferredMessageTestRpcAndNetworkVariableComponent.ClientInstances.Clear();
+            DeferredMessageTestRpcComponent.ClientInstances.Clear();
+            DeferredMessageTestNetworkVariableComponent.ClientInstances.Clear();
+            m_SkipAddingPrefabsToClient = false;
             // Host is irrelevant, messages don't get sent to the host "client"
             m_UseHost = false;
-
-            m_RpcPrefab = new GameObject("Object With RPC");
-            var networkObject = m_RpcPrefab.AddComponent<NetworkObject>();
-            m_RpcPrefab.AddComponent<DeferredMessageTestRpcComponent>();
-
-            // Make it a prefab
-            NetcodeIntegrationTestHelpers.MakeNetworkObjectTestPrefab(networkObject);
-
-            m_NetworkVariablePrefab = new GameObject("Object With NetworkVariable");
-            networkObject = m_NetworkVariablePrefab.AddComponent<NetworkObject>();
-            m_NetworkVariablePrefab.AddComponent<DeferredMessageTestNetworkVariableComponent>();
-
-            // Make it a prefab
-            NetcodeIntegrationTestHelpers.MakeNetworkObjectTestPrefab(networkObject);
-
-            m_RpcAndNetworkVariablePrefab = new GameObject("Object With NetworkVariable And RPC");
-            networkObject = m_RpcAndNetworkVariablePrefab.AddComponent<NetworkObject>();
-            m_RpcAndNetworkVariablePrefab.AddComponent<DeferredMessageTestRpcAndNetworkVariableComponent>();
-
-            // Make it a prefab
-            NetcodeIntegrationTestHelpers.MakeNetworkObjectTestPrefab(networkObject);
 
             // Replace the IDeferredMessageManager component with our test one in the component factory
             ComponentFactory.Register<IDeferredMessageManager>(networkManager => new TestDeferredMessageManager(networkManager));
@@ -269,13 +294,54 @@ namespace Unity.Netcode.RuntimeTests
 
         protected override void OnServerAndClientsCreated()
         {
-            m_ServerNetworkManager.AddNetworkPrefab(m_RpcPrefab);
-            m_ServerNetworkManager.AddNetworkPrefab(m_NetworkVariablePrefab);
-            m_ServerNetworkManager.AddNetworkPrefab(m_RpcAndNetworkVariablePrefab);
+            // Note: This is where prefabs should be created
+            m_RpcPrefab = CreateNetworkObjectPrefab("Object With RPC");
+            var networkObject = m_RpcPrefab.GetComponent<NetworkObject>();
+            m_RpcPrefab.AddComponent<DeferredMessageTestRpcComponent>();
+
+            m_NetworkVariablePrefab = CreateNetworkObjectPrefab("Object With NetworkVariable");
+            networkObject = m_NetworkVariablePrefab.GetComponent<NetworkObject>();
+            m_NetworkVariablePrefab.AddComponent<DeferredMessageTestNetworkVariableComponent>();
+
+            m_RpcAndNetworkVariablePrefab = CreateNetworkObjectPrefab("Object With NetworkVariable And RPC");
+            networkObject = m_RpcAndNetworkVariablePrefab.GetComponent<NetworkObject>();
+            m_RpcAndNetworkVariablePrefab.AddComponent<DeferredMessageTestRpcAndNetworkVariableComponent>();
+
             m_ServerNetworkManager.NetworkConfig.ForceSamePrefabs = false;
-            foreach (var client in m_ClientNetworkManagers)
+
+        }
+
+        private bool m_SkipAddingPrefabsToClient = false;
+
+        private void AddPrefabsToClient(NetworkManager networkManager)
+        {
+            networkManager.AddNetworkPrefab(m_RpcPrefab);
+            networkManager.AddNetworkPrefab(m_NetworkVariablePrefab);
+            networkManager.AddNetworkPrefab(m_RpcAndNetworkVariablePrefab);
+        }
+
+        protected override void OnNewClientCreated(NetworkManager networkManager)
+        {
+            networkManager.NetworkConfig.ForceSamePrefabs = false;
+            if (!m_SkipAddingPrefabsToClient)
             {
-                client.NetworkConfig.ForceSamePrefabs = false;
+                AddPrefabsToClient(networkManager);
+            }
+
+            base.OnNewClientCreated(networkManager);
+        }
+
+        private IEnumerator SpawnClients(bool clearTestDeferredMessageManagerCallFlags = true)
+        {
+            for (int i = 0; i < m_NumberOfClientsToLateJoin; i++)
+            {
+                // Create and join client
+                yield return CreateAndStartNewClient();
+            }
+
+            if (clearTestDeferredMessageManagerCallFlags)
+            {
+                ClearTestDeferredMessageManagerCallFlags();
             }
         }
 
@@ -299,21 +365,6 @@ namespace Unity.Netcode.RuntimeTests
                 var catcher = new SpawnCatcher();
                 m_ClientSpawnCatchers.Add(catcher);
                 client.MessagingSystem.Hook(catcher);
-            }
-        }
-
-        private void RegisterClientPrefabs(bool clearTestDeferredMessageManagerCallFlags = true)
-        {
-            foreach (var client in m_ClientNetworkManagers)
-            {
-                client.AddNetworkPrefab(m_RpcPrefab);
-                client.AddNetworkPrefab(m_NetworkVariablePrefab);
-                client.AddNetworkPrefab(m_RpcAndNetworkVariablePrefab);
-            }
-
-            if (clearTestDeferredMessageManagerCallFlags)
-            {
-                ClearTestDeferredMessageManagerCallFlags();
             }
         }
 
@@ -433,6 +484,21 @@ namespace Unity.Netcode.RuntimeTests
             return waiters;
         }
 
+        private List<IEnumerator> WaitForAllClientsToReceive<TFirstMessage, TSecondMessage, TThirdMessage>()
+        where TFirstMessage : INetworkMessage
+        where TSecondMessage : INetworkMessage
+        where TThirdMessage : INetworkMessage
+        {
+            var waiters = new List<IEnumerator>();
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                waiters.Add(NetcodeIntegrationTestHelpers.WaitForMessageOfTypeReceived<TFirstMessage>(client));
+                waiters.Add(NetcodeIntegrationTestHelpers.WaitForMessageOfTypeReceived<TSecondMessage>(client));
+                waiters.Add(NetcodeIntegrationTestHelpers.WaitForMessageOfTypeReceived<TThirdMessage>(client));
+            }
+            return waiters;
+        }
+
         private List<IEnumerator> WaitForAllClientsToReceive<TFirstMessage, TSecondMessage, TThirdMessage, TFourthMessage>()
             where TFirstMessage : INetworkMessage
             where TSecondMessage : INetworkMessage
@@ -454,7 +520,7 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenAnRpcArrivesBeforeASpawnArrives_ItIsDeferred()
         {
-            RegisterClientPrefabs();
+            yield return SpawnClients();
             CatchSpawns();
             var serverObject = Object.Instantiate(m_RpcPrefab);
             serverObject.GetComponent<NetworkObject>().NetworkManagerOwner = m_ServerNetworkManager;
@@ -477,7 +543,7 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenADespawnArrivesBeforeASpawnArrives_ItIsDeferred()
         {
-            RegisterClientPrefabs();
+            yield return SpawnClients();
             CatchSpawns();
             var serverObject = Object.Instantiate(m_RpcPrefab);
             serverObject.GetComponent<NetworkObject>().NetworkManagerOwner = m_ServerNetworkManager;
@@ -500,7 +566,7 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenAChangeOwnershipMessageArrivesBeforeASpawnArrives_ItIsDeferred()
         {
-            RegisterClientPrefabs();
+            yield return SpawnClients();
             CatchSpawns();
             var serverObject = Object.Instantiate(m_RpcPrefab);
             serverObject.GetComponent<NetworkObject>().NetworkManagerOwner = m_ServerNetworkManager;
@@ -521,12 +587,13 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenANetworkVariableDeltaMessageArrivesBeforeASpawnArrives_ItIsDeferred()
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
 
             // Have to start these before spawning here because spawning sends a NetworkVariableDeltaMessage, too
             // Depending on timing, if we start this after spawning, we may end up missing the first one.
-            var waiters = WaitForAllClientsToReceive<NetworkVariableDeltaMessage, NetworkVariableDeltaMessage>();
+            var waiters = WaitForAllClientsToReceive<NetworkVariableDeltaMessage>();
             var coroutines = StartMultiple(waiters);
 
             var serverObject = Object.Instantiate(m_NetworkVariablePrefab);
@@ -545,13 +612,16 @@ namespace Unity.Netcode.RuntimeTests
                 Assert.IsFalse(manager.ProcessTriggersCalled);
                 // TODO: Network Variables generate an extra message immediately at spawn for some reason...
                 // Seems like a bug since the network variable data is in the spawn message already.
-                AssertSpawnTriggerCountForObject(manager, serverObject, 2);
+                AssertSpawnTriggerCountForObject(manager, serverObject, 1);
             }
         }
 
         [UnityTest]
+        //[Ignore("Disabling this temporarily until it is migrated into new integration test.")]
         public IEnumerator WhenASpawnMessageArrivesBeforeThePrefabIsAvailable_ItIsDeferred()
         {
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             var serverObject = Object.Instantiate(m_RpcPrefab);
             serverObject.GetComponent<NetworkObject>().NetworkManagerOwner = m_ServerNetworkManager;
             serverObject.GetComponent<NetworkObject>().Spawn();
@@ -631,7 +701,27 @@ namespace Unity.Netcode.RuntimeTests
         public IEnumerator WhenANetworkVariableDeltaMessageIsDeferred_ItIsProcessedOnSpawn()
         {
             yield return WhenANetworkVariableDeltaMessageArrivesBeforeASpawnArrives_ItIsDeferred();
+
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                AddPrefabsToClient(client);
+            }
+
             ReleaseSpawns();
+
+            // Wait for the clients to spawn the NetworkObjects
+            bool HaveAllClientsSpawned()
+            {
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    if (!DeferredMessageTestNetworkVariableComponent.ClientInstances.Contains(client.LocalClientId))
+                    {
+                        return false;
+                    }
+                }
+                return true;
+            }
+            yield return WaitForConditionOrTimeOut(HaveAllClientsSpawned);
 
             foreach (var client in m_ClientNetworkManagers)
             {
@@ -647,9 +737,31 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenASpawnMessageIsDeferred_ItIsProcessedOnAddPrefab()
         {
+            // This will prevent spawned clients from adding prefabs
+            m_SkipAddingPrefabsToClient = true;
             yield return WhenASpawnMessageArrivesBeforeThePrefabIsAvailable_ItIsDeferred();
-            RegisterClientPrefabs(false);
 
+            // Now add the prefabs
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                AddPrefabsToClient(client);
+            }
+
+            // Wait for the clients to spawn the NetworkObjects
+            bool HaveAllClientsSpawned()
+            {
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    if (!DeferredMessageTestRpcComponent.ClientInstances.Contains(client.LocalClientId))
+                    {
+                        return false;
+                    }
+                }
+                return true;
+            }
+            yield return WaitForConditionOrTimeOut(HaveAllClientsSpawned);
+
+            // Validate this test
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
@@ -664,12 +776,12 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenMultipleSpawnTriggeredMessagesAreDeferred_TheyAreAllProcessedOnSpawn()
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
 
-            // Have to start these before spawning here because spawning sends a NetworkVariableDeltaMessage, too
-            // Depending on timing, if we start this after spawning, we may end up missing the first one.
-            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
+            // We wait for the RPC message, the NetworkVariableDeltaMessage, and the ChangeOwnershipMessage messages
+            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
             var coroutines = StartMultiple(waiters);
 
             var serverObject = Object.Instantiate(m_RpcAndNetworkVariablePrefab);
@@ -679,6 +791,9 @@ namespace Unity.Netcode.RuntimeTests
 
             serverObject.GetComponent<DeferredMessageTestRpcAndNetworkVariableComponent>().SendTestClientRpc();
             serverObject.GetComponent<DeferredMessageTestRpcAndNetworkVariableComponent>().TestNetworkVariable.Value = 1;
+            // TODO: Remove this if we figure out how to work around the NetworkVariableDeltaMessage.Serialized issue at line 59
+            // Otherwise, we have to wait for at least 1 tick for the NetworkVariableDeltaMessage to be generated before changing ownership
+            yield return s_DefaultWaitForTick;
             serverObject.GetComponent<NetworkObject>().ChangeOwnership(m_ClientNetworkManagers[0].LocalClientId);
 
             // Should be received in order so we'll wait for the last one.
@@ -690,29 +805,49 @@ namespace Unity.Netcode.RuntimeTests
                 Assert.IsTrue(manager.DeferMessageCalled);
                 Assert.IsFalse(manager.ProcessTriggersCalled);
 
-                Assert.AreEqual(4, manager.DeferredMessageCountTotal());
-                Assert.AreEqual(4, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
-                Assert.AreEqual(4, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, serverObject.GetComponent<NetworkObject>().NetworkObjectId));
+                Assert.AreEqual(3, manager.DeferredMessageCountTotal());
+                Assert.AreEqual(3, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
+                Assert.AreEqual(3, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, serverObject.GetComponent<NetworkObject>().NetworkObjectId));
                 Assert.AreEqual(0, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnAddPrefab));
+                AddPrefabsToClient(client);
             }
+
             ReleaseSpawns();
 
+            // Wait for the clients to spawn the NetworkObjects
+            bool HaveAllClientsSpawned()
+            {
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    if (!DeferredMessageTestRpcAndNetworkVariableComponent.ClientInstances.Contains(client.LocalClientId))
+                    {
+                        return false;
+                    }
+                }
+                return true;
+            }
+            yield return WaitForConditionOrTimeOut(HaveAllClientsSpawned);
+            yield return new WaitForSeconds(0.1f);
+
+            // Validate the spawned objects
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
-                Assert.IsTrue(manager.ProcessTriggersCalled);
-                Assert.AreEqual(0, manager.DeferredMessageCountTotal());
+                Assert.IsTrue(manager.ProcessTriggersCalled, "Process triggers were not called!");
+                Assert.AreEqual(0, manager.DeferredMessageCountTotal(), $"Deferred message count ({manager.DeferredMessageCountTotal()}) is not zero!");
 
                 var component = GetComponentForClient<DeferredMessageTestRpcAndNetworkVariableComponent>(client.LocalClientId);
-                Assert.IsTrue(component.ClientRpcCalled);
-                Assert.AreEqual(1, component.TestNetworkVariable.Value);
-                Assert.AreEqual(m_ClientNetworkManagers[0].LocalClientId, component.OwnerClientId);
+                Assert.IsTrue(component.ClientRpcCalled, "Client RPC was not called!");
+                Assert.AreEqual(1, component.TestNetworkVariable.Value, $"Test {nameof(NetworkVariable<int>)} ({component.TestNetworkVariable.Value}) does not equal 1!");
+                Assert.AreEqual(m_ClientNetworkManagers[0].LocalClientId, component.OwnerClientId, $"{component.name} owner id ({component.OwnerClientId}) does not equal first client id ({m_ClientNetworkManagers[0].LocalClientId})");
             }
         }
 
         [UnityTest]
         public IEnumerator WhenMultipleAddPrefabTriggeredMessagesAreDeferred_TheyAreAllProcessedOnAddNetworkPrefab()
         {
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             var serverObject = Object.Instantiate(m_RpcPrefab);
             serverObject.GetComponent<NetworkObject>().NetworkManagerOwner = m_ServerNetworkManager;
             serverObject.GetComponent<NetworkObject>().Spawn();
@@ -733,9 +868,23 @@ namespace Unity.Netcode.RuntimeTests
                 Assert.AreEqual(0, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
                 Assert.AreEqual(2, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnAddPrefab));
                 Assert.AreEqual(2, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnAddPrefab, serverObject.GetComponent<NetworkObject>().GlobalObjectIdHash));
+                AddPrefabsToClient(client);
             }
 
-            RegisterClientPrefabs(false);
+            // Wait for the clients to spawn the NetworkObjects
+            bool HaveAllClientsSpawned()
+            {
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    if (!DeferredMessageTestRpcComponent.ClientInstances.Contains(client.LocalClientId))
+                    {
+                        return false;
+                    }
+                }
+                return true;
+            }
+            yield return WaitForConditionOrTimeOut(HaveAllClientsSpawned);
+
 
             foreach (var client in m_ClientNetworkManagers)
             {
@@ -769,8 +918,9 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenSpawnTriggeredMessagesAreDeferredBeforeThePrefabIsAdded_AddingThePrefabCausesThemToBeProcessed()
         {
-            // Because we're not waiting for the client to receive the spawn before we change the network variable value,
-            // there's only one NetworkVariableDeltaMessage this time.
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
+
             var waiters = WaitForAllClientsToReceive<CreateObjectMessage, ClientRpcMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
             var coroutines = StartMultiple(waiters);
 
@@ -780,10 +930,14 @@ namespace Unity.Netcode.RuntimeTests
 
             serverObject.GetComponent<DeferredMessageTestRpcAndNetworkVariableComponent>().SendTestClientRpc();
             serverObject.GetComponent<DeferredMessageTestRpcAndNetworkVariableComponent>().TestNetworkVariable.Value = 1;
+            // TODO: Remove this if we figure out how to work around the NetworkVariableDeltaMessage.Serialized issue at line 59
+            // Otherwise, we have to wait for at least 1 tick for the NetworkVariableDeltaMessage to be generated before changing ownership
+            yield return s_DefaultWaitForTick;
             serverObject.GetComponent<NetworkObject>().ChangeOwnership(m_ClientNetworkManagers[0].LocalClientId);
 
             yield return WaitMultiple(coroutines);
 
+            // Validate messages are deferred and pending
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
@@ -794,10 +948,26 @@ namespace Unity.Netcode.RuntimeTests
                 Assert.AreEqual(3, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, serverObject.GetComponent<NetworkObject>().NetworkObjectId));
                 Assert.AreEqual(1, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnAddPrefab));
                 Assert.AreEqual(1, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnAddPrefab, serverObject.GetComponent<NetworkObject>().GlobalObjectIdHash));
+                AddPrefabsToClient(client);
             }
 
-            RegisterClientPrefabs(false);
+            // Wait for the clients to spawn the NetworkObjects
+            bool HaveAllClientsSpawned()
+            {
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    if (!DeferredMessageTestRpcAndNetworkVariableComponent.ClientInstances.Contains(client.LocalClientId))
+                    {
+                        return false;
+                    }
+                }
+                return true;
+            }
+            yield return WaitForConditionOrTimeOut(HaveAllClientsSpawned);
 
+            yield return new WaitForSeconds(0.1f);
+
+            // Validate the test
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
@@ -813,10 +983,10 @@ namespace Unity.Netcode.RuntimeTests
         }
 
         [UnityTest]
-        [Ignore("This test is unstable (MTT-4146)")]
         public IEnumerator WhenAMessageIsDeferredForMoreThanTheConfiguredTime_ItIsRemoved([Values(1, 2, 3)] int timeout)
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
             foreach (var client in m_ClientNetworkManagers)
             {
@@ -870,23 +1040,30 @@ namespace Unity.Netcode.RuntimeTests
 
             yield return new WaitForSeconds(timeout + 0.1f);
 
-            Assert.AreEqual(NumberOfClients, purgeCount);
-            foreach (var client in m_ClientNetworkManagers)
+            bool HaveAllClientsPurged()
             {
-                var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
-                Assert.AreEqual(0, manager.DeferredMessageCountTotal());
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
+                    if (manager.DeferredMessageCountTotal() != 0)
+                    {
+                        return false;
+                    }
+                }
+                return true;
             }
+
+            yield return WaitForConditionOrTimeOut(HaveAllClientsPurged);
+            AssertOnTimeout("Timed out waiting for all clients to purge their deferred messages!");
         }
 
         [UnityTest]
-        [Ignore("This test is unstable on standalones")]
         public IEnumerator WhenMultipleMessagesForTheSameObjectAreDeferredForMoreThanTheConfiguredTime_TheyAreAllRemoved([Values(1, 2, 3)] int timeout)
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
-            // Have to start these before spawning here because spawning sends a NetworkVariableDeltaMessage, too
-            // Depending on timing, if we start this after spawning, we may end up missing the first one.
-            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
+            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
             var coroutines = StartMultiple(waiters);
 
             foreach (var client in m_ClientNetworkManagers)
@@ -932,35 +1109,42 @@ namespace Unity.Netcode.RuntimeTests
                     ++purgeCount;
                     var elapsed = Time.realtimeSinceStartup - start;
                     Assert.GreaterOrEqual(elapsed, timeout - 0.05f);
-                    Assert.AreEqual(4, manager.DeferredMessageCountTotal());
-                    Assert.AreEqual(4, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
-                    Assert.AreEqual(4, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, key));
+                    Assert.AreEqual(3, manager.DeferredMessageCountTotal());
+                    Assert.AreEqual(3, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
+                    Assert.AreEqual(3, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, key));
                     Assert.AreEqual(serverObject.GetComponent<NetworkObject>().NetworkObjectId, key);
                 };
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
                 manager.OnBeforePurge = beforePurge;
             }
 
-            yield return new WaitForSeconds(timeout + 0.1f);
-
-            Assert.AreEqual(NumberOfClients, purgeCount);
-            foreach (var client in m_ClientNetworkManagers)
+            bool HaveAllClientsPurged()
             {
-                var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
-                Assert.AreEqual(0, manager.DeferredMessageCountTotal());
+                foreach (var client in m_ClientNetworkManagers)
+                {
+                    var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
+                    if (manager.DeferredMessageCountTotal() != 0)
+                    {
+                        return false;
+                    }
+                }
+                return true;
             }
+
+            yield return WaitForConditionOrTimeOut(HaveAllClientsPurged);
+            AssertOnTimeout("Timed out waiting for all clients to purge their deferred messages!");
         }
 
         [UnityTest]
         public IEnumerator WhenMultipleMessagesForDifferentObjectsAreDeferredForMoreThanTheConfiguredTime_TheyAreAllRemoved([Values(1, 2, 3)] int timeout)
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
 
-            // Have to start these before spawning here because spawning sends a NetworkVariableDeltaMessage, too
-            // Depending on timing, if we start this after spawning, we may end up missing the first one.
-            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
-            waiters.AddRange(WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>());
+            // Since there are two unique objects we need to look for two sets of messages
+            var waiters = WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>();
+            waiters.AddRange(WaitForAllClientsToReceive<ClientRpcMessage, NetworkVariableDeltaMessage, ChangeOwnershipMessage>());
             var coroutines = StartMultiple(waiters);
 
             foreach (var client in m_ClientNetworkManagers)
@@ -1004,6 +1188,7 @@ namespace Unity.Netcode.RuntimeTests
 
             foreach (var unused in m_ClientNetworkManagers)
             {
+
                 LogAssert.Expect(LogType.Warning, $"[Netcode] Deferred messages were received for a trigger of type {IDeferredMessageManager.TriggerType.OnSpawn} with key {serverObject.GetComponent<NetworkObject>().NetworkObjectId}, but that trigger was not received within within {timeout} second(s).");
                 LogAssert.Expect(LogType.Warning, $"[Netcode] Deferred messages were received for a trigger of type {IDeferredMessageManager.TriggerType.OnSpawn} with key {serverObject2.GetComponent<NetworkObject>().NetworkObjectId}, but that trigger was not received within within {timeout} second(s).");
             }
@@ -1011,7 +1196,7 @@ namespace Unity.Netcode.RuntimeTests
             int purgeCount = 0;
             foreach (var client in m_ClientNetworkManagers)
             {
-                var remainingMessagesTotalThisClient = 8;
+                var remainingMessagesTotalThisClient = 6;
                 TestDeferredMessageManager.BeforePurgeDelegate beforePurge = (manager, key) =>
                 {
                     ++purgeCount;
@@ -1019,16 +1204,19 @@ namespace Unity.Netcode.RuntimeTests
                     Assert.GreaterOrEqual(elapsed, timeout - 0.05f);
                     Assert.AreEqual(remainingMessagesTotalThisClient, manager.DeferredMessageCountTotal());
                     Assert.AreEqual(remainingMessagesTotalThisClient, manager.DeferredMessageCountForType(IDeferredMessageManager.TriggerType.OnSpawn));
-                    Assert.AreEqual(4, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, key));
-                    remainingMessagesTotalThisClient -= 4;
+                    Assert.AreEqual(3, manager.DeferredMessageCountForKey(IDeferredMessageManager.TriggerType.OnSpawn, key));
+                    remainingMessagesTotalThisClient -= 3;
                 };
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
                 manager.OnBeforePurge = beforePurge;
             }
 
             yield return new WaitForSeconds(timeout + 0.1f);
-
-            Assert.AreEqual(NumberOfClients * 2, purgeCount);
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                AddPrefabsToClient(client);
+            }
+            Assert.AreEqual(m_NumberOfClientsToLateJoin * 2, purgeCount);
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
@@ -1039,7 +1227,8 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenADeferredMessageIsRemoved_OtherMessagesForSameObjectAreRemoved([Values(1, 2, 3)] int timeout)
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
             foreach (var client in m_ClientNetworkManagers)
             {
@@ -1112,9 +1301,14 @@ namespace Unity.Netcode.RuntimeTests
                 manager.OnBeforePurge = beforePurge;
             }
 
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                AddPrefabsToClient(client);
+            }
+
             yield return new WaitForSeconds(0.6f);
 
-            Assert.AreEqual(NumberOfClients, purgeCount);
+            Assert.AreEqual(m_NumberOfClientsToLateJoin, purgeCount);
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;
@@ -1125,7 +1319,8 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator WhenADeferredMessageIsRemoved_OtherMessagesForDifferentObjectsAreNotRemoved([Values(1, 2, 3)] int timeout)
         {
-            RegisterClientPrefabs();
+            m_SkipAddingPrefabsToClient = true;
+            yield return SpawnClients();
             CatchSpawns();
             foreach (var client in m_ClientNetworkManagers)
             {
@@ -1206,9 +1401,14 @@ namespace Unity.Netcode.RuntimeTests
                 manager.OnBeforePurge = beforePurge;
             }
 
+            foreach (var client in m_ClientNetworkManagers)
+            {
+                AddPrefabsToClient(client);
+            }
+
             yield return new WaitForSeconds(0.6f);
 
-            Assert.AreEqual(NumberOfClients, purgeCount);
+            Assert.AreEqual(m_NumberOfClientsToLateJoin, purgeCount);
             foreach (var client in m_ClientNetworkManagers)
             {
                 var manager = (TestDeferredMessageManager)client.DeferredMessageManager;